Modularize into packaging_info.go
