+++ src/frontend/handlers.go
@@ -16,10 +16,8 @@ package main
 
 import (
 	"context"
-	"encoding/json"
 	"fmt"
 	"html/template"
-	"io/ioutil"
 	"math/rand"
 	"net"
 	"net/http"
@@ -537,42 +535,4 @@ func stringinSlice(slice []string, val string) bool {
 		}
 	}
 	return false
-}
-
-type PackagingInfo struct {
-	Weight float32 `json:"weight"`
-	Width  float32 `json:"width"`
-	Height float32 `json:"height"`
-	Depth  float32 `json:"depth"`
-}
-
-func httpGetPackagingInfo(productId string) (PackagingInfo, error) {
-	// Make the GET request
-	url := os.Getenv("PACKAGING_SERVICE_URL") + productId
-	fmt.Println("Requesting packaging info from URL: ", url)
-	resp, err := http.Get(url)
-	if err != nil {
-		return PackagingInfo{}, err
-	}
-	defer resp.Body.Close()
-
-	// Check the response status code
-	if resp.StatusCode != http.StatusOK {
-		return PackagingInfo{}, fmt.Errorf("Unexpected status code: %d", resp.StatusCode)
-	}
-
-	// Read the JSON response body
-	responseBody, err := ioutil.ReadAll(resp.Body)
-	if err != nil {
-		return PackagingInfo{}, err
-	}
-
-	// Decode the JSON response into a Post struct
-	var packagingInfo PackagingInfo
-	err = json.Unmarshal(responseBody, &packagingInfo)
-	if err != nil {
-		return PackagingInfo{}, err
-	}
-
-	return packagingInfo, nil
 }
